pyproject.toml
[tool.poetry]
name = "signalaibot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
semaphore-bot = "^0.16.0"
feedparser = "^6.0.10"
asks = "^3.0.0"
beautifulsoup4 = "^4.12.2"
pyyaml = "^6.0.1"
pydantic = "^2.4.2"
aiocache = "^0.12.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
watchfiles = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
